requirements.txt
cffi
cryptography
idna
pyasn1
pyasn1-modules
pycparser
pyOpenSSL
pytz
six
xkcdpass
#PyCryptodome
pycrypto
wheel
pytest
